requirements.txt
# VenueFlow AI — Python Dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-genai==1.14.0
python-dotenv==1.1.0
pydantic==2.11.1
pytest==8.3.5
